move moderator addition/remove into neutral territory
